# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface BlogPost {
  id: number;
  slug: string;
  title: string;
  excerpt: string;
  content: string;
  category: string;
  author: string;
  imageUrl: string;
  readTimeMinutes: number;
  publishedAt: string;
}

export interface PortfolioProject {
  id: number;
  title: string;
  category: string;
  description: string;
  imageUrl: string;
  location: string;
  clientName: string;
  completedAt: string;
}

export interface Service {
  id: number;
  slug: string;
  name: string;
  description: string;
  icon: string;
  priceFrom: number;
  duration: string;
  features: string[];
}

export interface Testimonial {
  id: number;
  customerName: string;
  role: string;
  quote: string;
  rating: number;
  imageUrl: string;
}

export interface FaqItem {
  id: number;
  question: string;
  answer: string;
  category: string;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  subject: string;
  message: string;
  createdAt: string;
}

export interface ContactMessageInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 1 */
  subject: string;
  /** @minLength 1 */
  message: string;
}

export interface NewsletterSubscriber {
  id: number;
  email: string;
  subscribedAt: string;
}

export interface NewsletterSubscriberInput {
  /** @minLength 3 */
  email: string;
}

export type AppointmentStatus = typeof AppointmentStatus[keyof typeof AppointmentStatus];


export const AppointmentStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
} as const;

export interface Appointment {
  id: number;
  name: string;
  email: string;
  phone: string;
  serviceName: string;
  preferredDate: string;
  preferredTime: string;
  /** @nullable */
  notes: string | null;
  status: AppointmentStatus;
  createdAt: string;
}

export interface AppointmentInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 1 */
  phone: string;
  /** @minLength 1 */
  serviceName: string;
  preferredDate: string;
  preferredTime: string;
  notes?: string;
}

export type AppointmentStatusUpdateStatus = typeof AppointmentStatusUpdateStatus[keyof typeof AppointmentStatusUpdateStatus];


export const AppointmentStatusUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
} as const;

export interface AppointmentStatusUpdate {
  status: AppointmentStatusUpdateStatus;
}

export interface TrafficSourceStat {
  source: string;
  visitors: number;
  percentage: number;
}

export interface TopPageStat {
  path: string;
  views: number;
  avgTimeSeconds: number;
}

export interface VisitorTrendPoint {
  date: string;
  visitors: number;
  pageViews: number;
}

export interface AnalyticsSummary {
  totalVisitors: number;
  totalPageViews: number;
  conversionRate: number;
  avgSessionMinutes: number;
  bounceRate: number;
  topPages: TopPageStat[];
  trafficBySource: TrafficSourceStat[];
  visitorsTrend: VisitorTrendPoint[];
}

export interface AdminSummary {
  totalAppointments: number;
  pendingAppointments: number;
  totalSubscribers: number;
  totalContactMessages: number;
  recentAppointments: Appointment[];
  recentContactMessages: ContactMessage[];
  recentSubscribers: NewsletterSubscriber[];
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}